requirements.txt
vobject

django-dept-classes>=0.5.0
django-dept-directory
django-dept-students
django-webcal

python-latex>=0.5.2
